+++ app/src/main/java/com/itisuniqueofficial/lockify/core/navigation/Screen.kt
@@ -11,5 +11,7 @@ sealed class Screen(val route: String) {
     object Settings : Screen("settings")
     object TriggerExclusions : Screen("trigger_exclusions")
     object AntiUninstall: Screen("anti_uninstall")
+    object Vault : Screen("vault")
+    object Stats : Screen("stats")
 }